refactor(tui): split stats_tab lower panels out; hold meta.rs at the cap
The 500-LoC gate was still red on two files, both mine, and neither was one
I had been trimming — I had been reducing files that turned out to be
allow-listed. Reproducing the CI check locally (parsing the allow-list
straight out of the workflow so it cannot drift) named the actual two:

* stats_tab.rs 414 -> 540, from this branch's PREFILL ingest tile and the
  dual-scale throughput chart.
* meta.rs 500 -> 502, from the request-sized ctx accumulator commit.

stats_tab.rs is a real split, not a trim: SEQUENCES & MEMORY, SPECULATION &
CACHE and the line_gauge they share move to stats_tab_panels.rs, following
the existing main_tab / main_tab_kernels sibling idiom. The cut is at a
genuine seam — everything moved renders the BOTTOM half of the tab, nothing
above it calls in, and the halves now share only App and the panel chrome.
326 + 237 lines.

meta.rs holds at 500 by compressing two of its own comments.

Adding allow-list entries was the wrong tool here: that list requires a
rationale AND a tracking issue per entry, which is the right bar and not one
a comment-verbosity problem should be allowed to clear.

2295 spark-server tests pass, including the stats-tab tile-width regression
tests that guard the very layout this moves.
